Move existing openapi GitHub release tests to WebhookIT
